Move Hero & Interstitial blocks to new directory structure.
